--- client/src/types/editButtonType.ts
@@ -1,3 +1,0 @@
-export interface disabeldProps {
-    disabled: boolean;
-}
