+++ client/src/types/variableType.ts
@@ -1,0 +1,7 @@
+export interface TitleTextProps {
+    titleText: string;
+}
+
+export interface disabeldProps {
+    disabled: boolean;
+}
